Add User

+++ Tosi_Back/src/main/java/com/ssafy/tosi/customTale/CustomTaleRepository.java
@@ -4,10 +4,10 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface CustomTaleRepository extends JpaRepository<CustomTale, Long> {
+public interface CustomTaleRepository extends JpaRepository<CustomTale, Integer> {
 
     // userId로 CustomTale 조회
-    List<CustomTale> findByUserId(Long userId);
+    List<CustomTale> findByUserId(Integer userId);
 
     // isPublic이 true인 CustomTale 조회
     List<CustomTale> findByIsPublic(boolean isPublic);
